move FailsafeTask into ClusterClient
