+++ connection/cluster/ClusterDatabaseManager.java
@@ -45,7 +45,7 @@ import static java.util.stream.Collectors.toMap;
 
 public class ClusterDatabaseManager implements DatabaseManager.Cluster {
 
-    private static final Logger LOG = LoggerFactory.getLogger(FailsafeTask.class);
+    private static final Logger LOG = LoggerFactory.getLogger(ClusterClient.FailsafeTask.class);
 
     private final Map<String, TypeDBDatabaseManagerImpl> databaseMgrs;
     private final ClusterClient client;
@@ -99,18 +99,12 @@ public class ClusterDatabaseManager implements DatabaseManager.Cluster {
     }
 
     private <RESULT> RESULT failsafeTask(String name, BiFunction<ClusterServerStub, TypeDBDatabaseManagerImpl, RESULT> task) {
-        FailsafeTask<RESULT> failsafeTask = new FailsafeTask<RESULT>(client, name) {
+        ClusterClient.FailsafeTask<RESULT> failsafeTask = ClusterClient.createFailsafeTask(
+                client,
+                name,
+                replica -> task.apply(client.stub(replica.address()), client.clusterServerClient(replica.address()).databases())
+        );
 
-            @Override
-            RESULT run(ClusterDatabase.Replica replica) {
-                return task.apply(client.stub(replica.address()), client.clusterServerClient(replica.address()).databases());
-            }
-
-            @Override
-            RESULT rerun(ClusterDatabase.Replica replica) {
-                return run(replica);
-            }
-        };
         try {
             return failsafeTask.runAnyReplica();
         } catch (TypeDBClientException e) {
